nvmet: add abort list infrastructure
Each command entering the target is referenced by the sq->state_list.
This is needed to all allow the lookup of future nvme abort and
cancel commands. For now, we use a simple spinlock to protect this list.
Optimization of this list/lock will be considered in a future patch.

Modeled after the scsi soft target abort handling code.

Signed-off-by: Maurizio Lombardi <mlombard@redhat.com>