nvme: multipath: Implemented new iopolicy "queue-depth"
The round-robin path selector is inefficient in cases where there is a
difference in latency between multiple active optimized paths.  In the
presence of one or more high latency paths the round-robin selector
continues to the high latency path equally. This results in a bias
towards the highest latency path and can cause is significant decrease
in overall performance as IOs pile on the lowest latency path. This
problem is particularly accute with NVMe-oF controllers.

The queue-depth policy instead sends I/O requests down the path with the
least amount of requests in its request queue. Paths with lower latency
will clear requests more quickly and have less requests in their queues
compared to higher latency paths. The goal of this path selector is to
make more use of lower latency paths, which will bring down overall IO
latency.

Signed-off-by: Ewan D. Milne <emilne@redhat.com>
[jmeneghi: vairious improvements to address review comments]
Co-developed-by: John Meneghini <jmeneghi@redhat.com>
[tsong: patch initially developed by Thomas Song @ Pure Storage]
Co-developed-by: Thomas Song <tsong@purestorage.com>
Tested-by: Marco Patalano <mpatalan@redhat.com>
Reviewed-by: Randy Jennings <randyj@redhat.com>
Tested-by: Jyoti Rani <jani@purestorage.com>
Tested-by: John Meneghini <jmeneghi@redhat.com>