aacraid: reply queue mapping based on IRQ affinity again
Changed Subject (to be read as): Updated Patch for CPU Offlining Feature - Request for Testing and Feedback

Since then, John reached out to me in a separate email thread,
suggesting we consider giving it a shot with the upstream Linux kernel
as well.  I've since integrated, optimized, and thoroughly tested the
patch, and I'm now sharing it with you all.

This patch is based on the 6.9 kernel,  which has rolled back the patch
I previously submitted that introduced a regression.  (ref -
https://patches.linaro.org/project/linux-scsi/patch/20230519230834.27436-1-sagar.biradar@microchip.com/)
In other wods - this patch addresses all outstanding issues and should
provide a more stable solution.

The patch : To minimize the patch's exposure, I've implemented a new
modparam, "aac_cpu_offline_feature". By default, this feature is
disabled (set to 0). However, you can enable it during driver load with
the following command:

insmod ./aacraid.ko aac_cpu_offline_feature=1

Enabling this feature allows for CPU offlining but may result in some IO
performance drop.  I recommend enabling the modparam during the driver
load, as the relevant changes are part of the init routine.

Please apply this patch, test it on your systems, and let me know how it
works for you. Also, feel free to share this patch with the customer and
gather their feedback.  Any observations regarding the change in
performance would be great to have.

Once I hear from you and the end customer - I plan to proceed with the
upstream submission.  Thank you for your cooperation and support.

Signed-off-by: John Meneghini <jmeneghi@redhat.com>