scsi: st: clear pos_unknown when the por ua is expected
The change for commit 9604eea5bd3a ("scsi: st: Add third party poweron
reset handling") was introduced to handle the problem where an
unexpected power/on reset resulted in an undetected position lost
condition which caused lost data as the driver wrote a FM at the
beginning of the tape. This happened to a customer several times while
using an Amazon tape gateway. A connection loss condition in the fabric
resulted in the gateway tape device resetting itself and rewinding the
tape. The st driver swallowed the Unit Attention and wrote a file mark
at BOT.

To solve this problem we added code which detected the unexpected a
power on/reset Unit Attention to st_chk_results(). This correctly set
pos_unknown in the driver and prevented further access to the device
until the tape is re-positioned or rewound.

The problem with this solution is that it is catching POR Unit Attentions
that are expected as well as those that are unexpected. This results in
an unwanted EIO response to MTIOCGET.

The change for commit 3d882cca73be ("scsi: st: Fix input/output error on
empty drive reset") attempts to fix this problem by returning success in
flush_buffer() when ready != to ST_READY.  The problem with this
solution is that, once ST_READY is set, if pos_unknown remains
uncleared, MTIOGET can still return EIO. This is confusing as some tape
drives will set a POR UA when the system reboots or the driver is
reloaded.

To solve this problem we clear pos_unknown in test_ready() which is
deigned to flush out exected Unit Attentions, appropriately, unless
device->was_reset is valid.

Fixes: 3d882cca73be ("scsi: st: Fix input/output error on empty drive reset")
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219419
Signed-off-by: John Meneghini <jmeneghi@redhat.com>
Reviewed-by: Laurence Oberman <loberman@redhat.com>
Tested-by: Laurence Oberman <loberman@redhat.com>