scsi: st: don't clear pos_unknown after por ua in check_tape()
check_tape() is clearing pos_unknown after a poweron/reset UA and
beginning a new_session. This risks data corruption as a poweron reset
results in a position lost condition on the tape. Following a poweron
reset condition a MTIOCTOP operation that rewinds or reposition the tape
should clear pos_unknown.

Also be sure to clear was_reset when ever we clear pos_unknown. If we don't
then the code in st_chk_result() will turn on pos_unknown with any
future check condition, causing confusion.

Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219419
Signed-off-by: John Meneghini <jmeneghi@redhat.com>
Tested-by: John Meneghini <jmeneghi@redhat.com>