scsi: st: allow mtioget after por ua/reset"
The change for commit 9604eea5bd3a ("scsi: st: Add third party poweron
reset handling") was introduced to handle the problem where an
unexpected power/on reset resulted in an undetected position lost
condition which caused lost data as the driver flushed data and wrote a
file mark at the beginning of the tape.

To solve this problem code was added which detected the unexpected a
power on/reset Unit Attention in st_chk_results(). This correctly set
pos_unknown in the driver and prevented further access to the device
until the tape is re-positioned or rewound.

The problem with this solution is that it is catching POR Unit
Attentions that are expected as well as those that are unexpected. This
results in an unwanted EIO response to MTIOCGET.

The change for commit 3d882cca73be ("scsi: st: Fix input/output error on
empty drive reset") attempts to fix this problem by returning success in
flush_buffer() when ready != to ST_READY. However, once ST_READY is set,
if pos_unknown remains uncleared, MTIOGET can still return EIO. This is
confusing as some tape drives will set a POR UA when the system
reboots or the driver is reloaded.

To reproduce:

 # lsscsi -N -ig | grep tape
   [6:0:0:0]    tape    QUANTUM  ULTRIUM 4   U53F  /dev/st0   -  /dev/sg1
 # dd if=/dev/random count=1001024 of=/dev/nst0 &
 # sleep 5
 # sg_reset --target /dev/sg1
   dd: writing to '/dev/nst0': Input/output error
   dd: closing output file '/dev/nst0': Input/output error
   [1]+  Exit 1              dd if=/dev/random count=1001024 of=/dev/nst0
 # mt -f /dev/nst0 status
   /dev/nst0: Input/output error
 # mt -f /dev/nst0 status
   /dev/nst0: Input/output error
 # mt -f /dev/nst0 status
   /dev/nst0: Input/output error

Fixes: 3d882cca73be ("scsi: st: Fix input/output error on empty drive reset")
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219419
Signed-off-by: John Meneghini <jmeneghi@redhat.com>
Suggested-by: Kai Makisara <kai.makisara@kolumbus.fi>
Tested-by: John Meneghini <jmeneghi@redhat.com>