scsi: st: don't clear pos_unknown after por ua in test_ready()
check_tape() is beginning a new session and clearing pos_unknown when a
poweron/reset UA is encountered in test_ready(). This risks data
corruption as a poweron reset results in a position lost condition on
the tape. Following a poweron reset only a MTIOCTOP operation that
rewinds or reposition the tape should clear pos_unknown.

Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219419
Signed-off-by: John Meneghini <jmeneghi@redhat.com>
Tested-by: John Meneghini <jmeneghi@redhat.com>