scsi: st: Modify st.c to use the new scsi_error counters
Compare the stored values of por_ctr and new_media_ctr against
the values in the device struct. In case of mismatch, the
Unit Attention corresponding to the counter has happened.
This is a safeguard against another ULD catching the
Unit Attention sense data.

Remove use of the was_reset flag in struct scsi_device.

Signed-off-by: Kai Mäkisara <Kai.Makisara@kolumbus.fi>