scsi: scsi_error: Add counters for New Media and Power On/Reset UNIT ATTENTIONs
The purpose of the counters is to enable all ULDs attached to a
device to find out that a New Media or/and Power On/Reset Unit
Attentions has/have been set, even if another ULD catches the
Unit Attention as response to a SCSI command.

The ULDs can read the counters and see if the values have changed from
the previous check.

Signed-off-by: Kai Mäkisara <Kai.Makisara@kolumbus.fi>